cli: Optionally watch the task until it completes
This includes a dirty trick to be able to reuse the task-watching code
from /usr/bin/koji.

It's the best (or rather least horrible) I could find, given that it is
not part of the public Koji API.

But we should really work on exposing this as public API, so that
everybody can use it without tricks like this, or worse: copying it
over and over again.

At the very least, I've seen pyrpkg (or was it fedpkg?) copy-pasting
this.