builder: Make the mashed tree available outside of the buildroot
We bind mount the Koji topdir (usually /mnt/koji) inside the mock
chroot, and we do so read/write.

This is pretty risky business though, which makes this plugin slightly
worrisome.

In fact, this is the main reason why we had to implement this as a
plugin: upstream Koji developers really didn't want anything in core
Koji requiring the topdir to be read/write mounted inside the
buildroots.