cli: Merge the config and arguments
With this change, we can use the options the same way whether they come
from the config file, or from the command-line arguments.

Also, it allows the command-line arguments to override the values in the
config file, if an option exists in both.

This will all make sense soon, I promise.